Rename expectedOutputCalculators to outputDisplayFormatters
- Rename utility module to better reflect its purpose
- Update import statements across components
- Remove old test and implementation files
- Add new test and implementation files with clearer naming
