+++ apps/frontend/src/utils/outputDisplayFormatters.ts
